restructurado segun jerarquia :smirk:
