--- src/tests/PCI/Models/AddressTest.php
@@ -1,30 +1,0 @@
-<?php namespace Tests\PCI\Models;
-
-use PCI\Models\Address;
-use PCI\Models\Parish;
-use PCI\Models\Employee;
-use Tests\BaseTestCase;
-
-class AddressTest extends BaseTestCase
-{
-
-    public function testParish()
-    {
-        $this->mockBasicModelRelation(
-            Address::class,
-            'parish',
-            'belongsTo',
-            Parish::class
-        );
-    }
-
-    public function testUserDetails()
-    {
-        $this->mockBasicModelRelation(
-            Address::class,
-            'employee',
-            'hasMany',
-            Employee::class
-        );
-    }
-}
