+++ src/tests/PCI/Models/Item/SubCategory/SubCategoryTest.php
@@ -1,0 +1,31 @@
+<?php namespace Tests\PCI\Models\Item\SubCategory;
+
+use Mockery;
+use PCI\Models\Category;
+use PCI\Models\Item;
+use PCI\Models\SubCategory;
+use Tests\BaseTestCase;
+
+class SubCategoryTest extends BaseTestCase
+{
+
+    public function testCategory()
+    {
+        $this->mockBasicModelRelation(
+            SubCategory::class,
+            'category',
+            'belongsTo',
+            Category::class
+        );
+    }
+
+    public function testItems()
+    {
+        $this->mockBasicModelRelation(
+            SubCategory::class,
+            'items',
+            'hasMany',
+            Item::class
+        );
+    }
+}
