--- src/tests/PCI/Models/MakerTest.php
@@ -1,20 +1,0 @@
-<?php namespace Tests\PCI\Models;
-
-use Mockery;
-use PCI\Models\Item;
-use PCI\Models\Maker;
-use Tests\BaseTestCase;
-
-class MakerTest extends BaseTestCase
-{
-
-    public function testItems()
-    {
-        $this->mockBasicModelRelation(
-            Maker::class,
-            'items',
-            'hasMany',
-            Item::class
-        );
-    }
-}
