+++ src/tests/PCI/Models/Note/Type/NoteTypeTest.php
@@ -1,0 +1,19 @@
+<?php namespace Tests\PCI\Models\Note\Type;
+
+use PCI\Models\Note;
+use PCI\Models\NoteType;
+use Tests\BaseTestCase;
+
+class NoteTypeTest extends BaseTestCase
+{
+
+    public function testNotes()
+    {
+        $this->mockBasicModelRelation(
+            NoteType::class,
+            'notes',
+            'hasMany',
+            Note::class
+        );
+    }
+}
