--- src/tests/PCI/Models/NoteTypeTest.php
@@ -1,19 +1,0 @@
-<?php namespace Tests\PCI\Models;
-
-use PCI\Models\Note;
-use PCI\Models\NoteType;
-use Tests\BaseTestCase;
-
-class NoteTypeTest extends BaseTestCase
-{
-
-    public function testNotes()
-    {
-        $this->mockBasicModelRelation(
-            NoteType::class,
-            'notes',
-            'hasMany',
-            Note::class
-        );
-    }
-}
