--- src/tests/PCI/Models/PositionTest.php
@@ -1,20 +1,0 @@
-<?php namespace Tests\PCI\Models;
-
-use Mockery;
-use PCI\Models\Position;
-use PCI\Models\WorkDetail;
-use Tests\BaseTestCase;
-
-class PositionTest extends BaseTestCase
-{
-
-    public function testWorkDetails()
-    {
-        $this->mockBasicModelRelation(
-            Position::class,
-            'workDetails',
-            'hasMany',
-            WorkDetail::class
-        );
-    }
-}
