--- src/tests/PCI/Models/TownTest.php
@@ -1,31 +1,0 @@
-<?php namespace Tests\PCI\Models;
-
-use Mockery;
-use PCI\Models\Parish;
-use PCI\Models\State;
-use PCI\Models\Town;
-use Tests\BaseTestCase;
-
-class TownTest extends BaseTestCase
-{
-
-    public function testState()
-    {
-        $this->mockBasicModelRelation(
-            Town::class,
-            'state',
-            'belongsTo',
-            State::class
-        );
-    }
-
-    public function testParishes()
-    {
-        $this->mockBasicModelRelation(
-            Town::class,
-            'parishes',
-            'hasMany',
-            Parish::class
-        );
-    }
-}
