+++ src/tests/PCI/Models/User/Employee/Nationality/NationalityTest.php
@@ -1,0 +1,19 @@
+<?php namespace Tests\PCI\Models\User\Employee\Nationality;
+
+use Mockery;
+use PCI\Models\Nationality;
+use PCI\Models\Employee;
+use Tests\BaseTestCase;
+
+class NationalityTest extends BaseTestCase
+{
+    public function testUserDetails()
+    {
+        $this->mockBasicModelRelation(
+            Nationality::class,
+            'employee',
+            'hasMany',
+            Employee::class
+        );
+    }
+}
