+++ src/tests/PCI/Models/User/Employee/WorkDetails/WorkDetailTest.php
@@ -1,0 +1,42 @@
+<?php namespace Tests\PCI\Models\User\Employee\WorkDetails;
+
+use Mockery;
+use PCI\Models\Department;
+use PCI\Models\Employee;
+use PCI\Models\Position;
+use PCI\Models\WorkDetail;
+use Tests\BaseTestCase;
+
+class WorkDetailTest extends BaseTestCase
+{
+
+    public function testEmployee()
+    {
+        $this->mockBasicModelRelation(
+            WorkDetail::class,
+            'employee',
+            'belongsTo',
+            Employee::class
+        );
+    }
+
+    public function testPosition()
+    {
+        $this->mockBasicModelRelation(
+            WorkDetail::class,
+            'position',
+            'belongsTo',
+            Position::class
+        );
+    }
+
+    public function testDepartment()
+    {
+        $this->mockBasicModelRelation(
+            WorkDetail::class,
+            'department',
+            'belongsTo',
+            Department::class
+        );
+    }
+}
