+++ src/tests/PCI/Models/User/Petition/Type/PetitionTypeTest.php
@@ -1,0 +1,19 @@
+<?php namespace Tests\PCI\Models\User\Petition\Type;
+
+use PCI\Models\Petition;
+use PCI\Models\PetitionType;
+use Tests\BaseTestCase;
+
+class PetitionTypeTest extends BaseTestCase
+{
+
+    public function testPetitions()
+    {
+        $this->mockBasicModelRelation(
+            PetitionType::class,
+            'petitions',
+            'hasMany',
+            Petition::class
+        );
+    }
+}
