+++ src/tests/PCI/Models/User/Profile/ProfileTest.php
@@ -1,0 +1,19 @@
+<?php namespace Tests\PCI\Models\User\Profile;
+
+use PCI\Models\Profile;
+use PCI\Models\User;
+use Tests\BaseTestCase;
+
+class ProfileTest extends BaseTestCase
+{
+
+    public function testUsers()
+    {
+        $this->mockBasicModelRelation(
+            Profile::class,
+            'users',
+            'hasMany',
+            User::class
+        );
+    }
+}
